Issue #6: Create Query Rest
