+++ core/src/main/java/co/com/sofka/application/commands/UseCaseFactoryException.java
@@ -1,0 +1,12 @@
+package co.com.sofka.application.commands;
+
+public class UseCaseFactoryException extends RuntimeException {
+
+    public UseCaseFactoryException() {
+    }
+
+    public UseCaseFactoryException(String message) {
+        super(message);
+    }
+
+}
